requirements.txt
# ASR Core Dependencies
openai-whisper>=20240930
torch>=2.0.0
torchaudio>=2.0.0

# Alternative Whisper Implementations (for optimization)
faster-whisper>=0.10.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
sounddevice>=0.4.6
pydub>=0.25.1
noisereduce>=3.0.0

# Voice Activity Detection
silero-vad>=5.1

# HuggingFace Ecosystem (for fine-tuning)
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.7.0
evaluate>=0.4.1

# Evaluation Metrics
jiwer>=3.0.0

# UI Framework
streamlit>=1.28.0
streamlit-webrtc>=0.47.0

# Configuration Management
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0

# Logging and Monitoring
loguru>=0.7.0
rich>=13.7.0

# Development Tools
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Jupyter (for experimentation)
jupyter>=1.0.0
ipykernel>=6.26.0
ipywidgets>=8.1.0

